feat(spec): T3-02 golden parity between vectorized engine and whole-lot matcher
Unified execution spec (ashare_portfolio/golden, EXECUTION_SPEC_VERSION
1): the same signals run through the engine, the matcher in lot-free
continuous-share mode and the matcher in whole-lot mode.  Lot-free is
exactly identical to the engine (frictionless-rebalanced convention with
authoritative fills); lot-mode divergence decomposes per day into the
recorded integer-lot rounding, fee financing and cash residuals, with an
exact bookkeeping identity.  Coverage: suspension, limit up/down, T+1,
delisting, qfq adjustment, cash, fees, sell-blocked exits; stress: cost
0.5x/1x/2x, delayed execution, missing bars, extreme limit moves,
capital scales, lot sizes.  Capstone: T3-01 optimizer weights execute
through the spec with constraints respected.

Additive engine/schema changes: engine execution_delay, exact
BacktestResult.target_weights, broker lot_size (default unchanged),
shared order builder (ashare_trading/orders) now used by run_sim.

+++ ashare_trading/portfolio.py
@@ -14,8 +14,8 @@ from ashare_data.io_utils import atomic_write_json, read_json_safe
 class PositionState:
     ts_code: str
     name: str
-    quantity: int
-    available_quantity: int
+    quantity: float
+    available_quantity: float
     avg_cost: float
     last_price: float
     last_date: str
@@ -92,7 +92,7 @@ class SimulationPortfolio:
         # Atomic write: a crash mid-save can never leave a truncated state.
         atomic_write_json(self.state_path, payload)
 
-    def add_buy(self, ts_code: str, name: str, quantity: int, price: float, date: str) -> None:
+    def add_buy(self, ts_code: str, name: str, quantity: float, price: float, date: str) -> None:
         cost = quantity * price
         self.cash -= cost
         pos = self.positions.get(ts_code)
@@ -115,7 +115,7 @@ class SimulationPortfolio:
             pos.last_date = date
         self.trade_count += 1
 
-    def add_sell(self, ts_code: str, quantity: int, price: float, date: str) -> None:
+    def add_sell(self, ts_code: str, quantity: float, price: float, date: str) -> None:
         pos = self.positions.get(ts_code)
         if pos is None:
             return
